refactor: merge test_parser_backends.py into test_parser.py, remove redundant categorical test
