+++ tests/test_parser.py
@@ -1,16 +1,9 @@
-"""Core parser tests.
+"""Parser tests — API, engines, parity, registration, DuckDB load.
 
-Tests the new triplets.parser API, pandas.read_RDF delegation, find_all_xml,
-clean_ID, meta rows (Distribution/NamespaceMap), nodeID support, categorical
-(dictionary) encoding for memory, and interop notes.
-
-Uses minimal self-contained data + RealGrid when present (session scoped).
+Uses minimal_cim.xml (committed) + RealGrid when present.
 """
-import os
-from pathlib import Path
-
-import pandas as pd
 import pytest
+import pandas
 
 import triplets
 from triplets.parser import parse, find_all_xml, clean_ID, read_rdf
@@ -18,6 +11,20 @@ from triplets.parser import parse, find_all_xml, clean_ID, read_rdf
 REALGRID_ZIP = "test_data/TestConfigurations_packageCASv2.0/RealGrid/CGMES_v2.4.15_RealGridTestConfiguration_v2.zip"
 MINIMAL = "tests/data/minimal_cim.xml"
 
+try:
+    import pyarrow  # noqa: F401
+    HAS_PYARROW = True
+except ImportError:
+    HAS_PYARROW = False
+
+try:
+    from triplets.parser import cython_pugixml_arrow  # noqa: F401
+    HAS_CYTHON_PUGIXML_ARROW = True
+except Exception:
+    HAS_CYTHON_PUGIXML_ARROW = False
+
+
+# ── Utilities ───────────────────────────────────────────────────────────────
 
 def test_clean_ID():
     assert clean_ID("urn:uuid:abc-123") == "abc-123"
@@ -32,17 +39,44 @@ def test_find_all_xml_minimal():
     assert len(files) == 1
 
 
+# ── Registration ────────────────────────────────────────────────────────────
+
+class TestRegistration:
+    def test_pandas_read_RDF_registered(self):
+        assert hasattr(pandas, "read_RDF")
+        assert hasattr(pandas, "read_rdf")
+        assert callable(pandas.read_RDF)
+
+    def test_polars_read_rdf_registered(self):
+        polars = pytest.importorskip("polars")
+        assert hasattr(polars, "read_rdf")
+        assert callable(polars.read_rdf)
+
+
+# ── Parse API (parametrized across engines) ─────────────────────────────────
+
 def test_parse_minimal_basic(parser_engine):
-    """Basic parse via new API + read_RDF delegation."""
     df = parse(MINIMAL, engine=parser_engine)
     assert list(df.columns) == ["ID", "KEY", "VALUE", "INSTANCE_ID"]
     assert len(df) > 0
     assert "Distribution" in df["VALUE"].values
     assert "NamespaceMap" in df["VALUE"].values
-    # read_RDF alias
-    df2 = pd.read_RDF(MINIMAL)
-    assert len(df2) == len(df)
 
+
+def test_nodeid_support(parser_engine):
+    df = parse(MINIMAL, engine=parser_engine)
+    assert any("ConnectivityNode" in v for v in df[df["KEY"] == "Type"]["VALUE"])
+
+
+@pytest.mark.parametrize("use_cat", [True, False])
+def test_categorical_encoding(parser_engine, use_cat):
+    cols = ("INSTANCE_ID", "KEY") if use_cat else None
+    df = parse(MINIMAL, engine=parser_engine, return_type="pandas", categorical_columns=cols)
+    if use_cat:
+        assert df["KEY"].nunique() < 20
+
+
+# ── Return types ────────────────────────────────────────────────────────────
 
 def test_parse_returns_arrow_and_polars():
     pa = pytest.importorskip("pyarrow")
@@ -55,47 +89,89 @@ def test_parse_returns_arrow_and_polars():
     assert len(pdf) > 0
 
 
-def test_nodeid_support(parser_engine):
-    # The minimal_cim has a rdf:nodeID example
-    df = parse(MINIMAL, engine=parser_engine)
-    # Should have captured the nodeID object
-    # The ConnectivityNode with nodeID should appear
-    assert any("ConnectivityNode" in v for v in df[df["KEY"] == "Type"]["VALUE"])
+# ── Per-engine tests ────────────────────────────────────────────────────────
+
+class TestPythonLxmlPandas:
+    def test_basic_load(self):
+        df = parse(MINIMAL, engine="python_lxml_pandas")
+        assert isinstance(df, pandas.DataFrame)
+        assert len(df) > 0
+        types = set(df[df["KEY"] == "Type"]["VALUE"])
+        assert "Substation" in types
 
 
-@pytest.mark.parametrize("use_cat", [True, False])
-def test_categorical_encoding(parser_engine, use_cat):
-    """INSTANCE_ID and KEY should be categorized when requested."""
-    cols = ("INSTANCE_ID", "KEY") if use_cat else None
-    df = parse(MINIMAL, engine=parser_engine, return_type="pandas", categorical_columns=cols)
-    if use_cat:
-        key_card = df["KEY"].nunique()
-        assert key_card < 20, "KEY should have low cardinality"
+@pytest.mark.skipif(not HAS_PYARROW, reason="pyarrow not installed")
+class TestPythonLxmlArrow:
+    def test_basic_load(self):
+        df = parse(MINIMAL, engine="python_lxml_arrow")
+        assert isinstance(df, pandas.DataFrame)
+        assert len(df) > 0
+
+    def test_return_arrow(self):
+        import pyarrow as pa
+        t = parse(MINIMAL, engine="python_lxml_arrow", return_type="arrow")
+        assert isinstance(t, pa.Table)
+        assert t.num_rows > 0
 
 
-def test_realgrid_parse_and_meta(realgrid_data):
-    """Smoke on RealGrid (large, session scoped)."""
-    data = realgrid_data
-    assert len(data) > 1_000_000
-    assert "ACLineSegment" in set(data[data["KEY"] == "Type"]["VALUE"])
-    # Meta rows present
-    assert "Distribution" in data["VALUE"].values
+@pytest.mark.skipif(not HAS_CYTHON_PUGIXML_ARROW, reason="cython_pugixml_arrow not built")
+class TestCythonPugixmlArrow:
+    def test_basic_load(self):
+        df = parse(MINIMAL, engine="cython_pugixml_arrow")
+        assert isinstance(df, pandas.DataFrame)
+        assert len(df) > 0
 
+
+# ── Engine parity ───────────────────────────────────────────────────────────
+
+class TestParity:
+    def test_row_count_parity(self):
+        engines = ["python_lxml_pandas"]
+        if HAS_PYARROW:
+            engines.append("python_lxml_arrow")
+        if HAS_CYTHON_PUGIXML_ARROW:
+            engines.append("cython_pugixml_arrow")
+
+        results = {eng: parse(MINIMAL, engine=eng, categorical_columns=None) for eng in engines}
+        counts = [len(df) for df in results.values()]
+        assert all(c == counts[0] for c in counts), f"Row count mismatch: {dict(zip(results.keys(), counts))}"
+
+    def test_content_parity(self):
+        engines = ["python_lxml_pandas"]
+        if HAS_PYARROW:
+            engines.append("python_lxml_arrow")
+        if HAS_CYTHON_PUGIXML_ARROW:
+            engines.append("cython_pugixml_arrow")
+        if len(engines) < 2:
+            pytest.skip("Need at least 2 engines")
+
+        results = {}
+        for eng in engines:
+            df = parse(MINIMAL, engine=eng, categorical_columns=None).astype(str)
+            types = df[df["KEY"] == "Type"]
+            meta_ids = set(types[types["VALUE"].isin(["Distribution", "NamespaceMap"])]["ID"])
+            data_df = df[~df["ID"].isin(meta_ids)]
+            results[eng] = data_df[["KEY", "VALUE"]].sort_values(["KEY", "VALUE"]).reset_index(drop=True).values.tolist()
+
+        ref_name = engines[0]
+        for eng in engines[1:]:
+            assert results[eng] == results[ref_name], f"{eng} differs from {ref_name}"
+
+
+# ── DuckDB load ─────────────────────────────────────────────────────────────
 
 def test_duckdb_read_rdf():
-    """DuckDB con.read_rdf() loads data via Arrow zero-copy."""
     duckdb = pytest.importorskip("duckdb")
     import triplets
     data = duckdb.connect()
     rows = data.read_rdf([MINIMAL])
     assert rows == 17
-    td = data.types_dict()
-    assert "Substation" in td
+    assert "Substation" in data.types_dict()
 
 
-def test_read_rdf_with_categorical(realgrid_data):
-    df = read_rdf(REALGRID_ZIP, engine="cython_pugixml_arrow", categorical_columns=("INSTANCE_ID", "KEY"))
-    # KEY should be categorical-like (low unique)
-    assert df["KEY"].nunique() < 500
-    # INSTANCE_ID low unique (one per XML file inside the zip)
-    assert df["INSTANCE_ID"].nunique() < 20
+# ── RealGrid (large data, session scoped) ───────────────────────────────────
+
+def test_realgrid_parse_and_meta(realgrid_data):
+    assert len(realgrid_data) > 1_000_000
+    assert "ACLineSegment" in set(realgrid_data[realgrid_data["KEY"] == "Type"]["VALUE"])
+    assert "Distribution" in realgrid_data["VALUE"].values
